fix(create.sql): add 'data_inicio' to the UNIQUE key
Each 'oferecimento' has an unique 'docente', 'aluno', and 'disciplina'.
However, that does't take into consideration the start date, so two
'oferecimentos' in two different years could end up conflicting with
each other.

To solve that, 'data_inicio' was added as part of the UNIQUE key, so
that 'oferecimentos' in different dates count as different
'oferecimentos'