feat(web): implement web application
The web application allows the user to edit the 'teste' database from
the user 'testeusr' and password 'testeusr' at localhost:8080.

Basic instructions about how to compile and execute it are on the README
markdown file.

As of now there is no home page to link all the functionalities, see the
controllers to see which paths are available to make GET requests.

Most of the files were generated automatically by the lazy.py script I
wrote. Java is too verbose, and often repetitive... All of the
bothersome files in the src directory (except Application.java) can be
generated from a single dictionary representing the logical schema of
the database.

/ᐠ_ ꞈ _ᐟ\ฅ Laziness wins this time~